Move modal to separate components so we can reuse it in gridlist etc.
